Move code down to super class
to reduce redundancy

+++ src/test/hstreamprocessor_test.cpp
@@ -23,17 +23,17 @@ class HStreamProcessor_Test: public Test
         {
             private:
 
-                int* _data;
+                int8_t* _data;
 
             public:
 
-                TestReader(int* data):
+                TestReader(int8_t* data):
                     _data(data)
                 {}
 
                 int Read(int8_t* dest, size_t blocksize)
                 {
-                    memcpy(dest, _data, blocksize * sizeof(int));
+                    memcpy(dest, _data, blocksize * sizeof(int8_t));
                     return blocksize;
                 }
         };
@@ -42,19 +42,19 @@ class HStreamProcessor_Test: public Test
         {
             private:
 
-                int* _received;
+                int8_t* _received;
                 bool* _terminated;
 
             public:
 
-                TestWriter(int* received, bool* terminationToken):
+                TestWriter(int8_t* received, bool* terminationToken):
                     _received(received),
                     _terminated(terminationToken)
                 {}
 
                 int Write(int8_t* src, size_t blocksize)
                 {
-                    memcpy(_received, src, blocksize * sizeof(int));
+                    memcpy(_received, src, blocksize * sizeof(int8_t));
                     *_terminated = true;
                     return blocksize;
                 }
@@ -63,9 +63,9 @@ class HStreamProcessor_Test: public Test
         void test_read_write()
         {
             bool terminated = false;
-            int expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
-            int received[14];
-            memset(received, 0, 14 * sizeof(int));
+            int8_t expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
+            int8_t received[14];
+            memset(received, 0, 14 * sizeof(int8_t));
 
             TestReader rdr(expected);
             TestWriter wr(received, &terminated);
@@ -73,6 +73,6 @@ class HStreamProcessor_Test: public Test
 
             proc.Run();
 
-            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int)), 0);
+            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int8_t)), 0);
         }
 } hstreamprocessor_test;
